backend/Authentication/serializers.py
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer

class CustomToken_Serializer(TokenObtainPairSerializer):
    def validate(self, attrs):
        data = super().validate(attrs)

        # Add custom claims
        data['username'] = self.user.username
        data['email'] = self.user.email
        data['first_name'] = self.user.first_name
        data['last_name'] = self.user.last_name

        return data